beta version

+++ database/migrations/2023_12_10_125851_moodle_task.php
@@ -13,10 +13,11 @@ return new class extends Migration
     {
         Schema::create('moodle_task', function (Blueprint $table) {
             $table->id();
-            $table->integer('link_id');
+            $table->bigInteger('link_id');
             $table->string('name');
             $table->string('type');
             $table->bigInteger('course_id', false, true);
+            
             $table->foreign('course_id')->references('id')->on('moodle_course');
         });
     }
